Convert js to ts
