+++ frontend/src/App.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+		<p>Hello</p>
+      </header>
+    </div>
+  );
+}
+
+export default App;
